\name{comparative_scan_rtfbs.ms}
\alias{comparative_scan_rtfbs.ms}
\title{
comparative_scan_rtfbs
}
\description{
}
\usage{
comparative_scan_rtfbs.ms(pwm, 
      positive.ms, negative.ms, 
      background.ms, 
      background.mm, 
      fdr = 0.1, 
      threshold = NA, 
      calc.empirical.pvalue = FALSE)
}
\arguments{
  \item{pwm}{}
  \item{positive.ms}{}
  \item{negative.ms}{}
  \item{background.ms}{}
  \item{background.mm}{}
  \item{fdr}{}
  \item{threshold}{}
  \item{calc.empirical.pvalue}{}
}
\details{
}
\value{
}
\seealso{
}
\examples{
}
\keyword{internal}
