\name{print.tfbs.finding}
\alias{print.tfbs.finding}
\title{
Print scanning result of TF sites.
}
\description{
Print scanning result of TF sites.
}
\usage{
\method{print}{tfbs.finding}(x, \dots)
}
\arguments{
  \item{x}{The result obtained by \code{\link{tfbs.scanTFsite}}.}
  \item{\dots}{Additional arguments affecting the print produced.}
}
\details{
This function shows a brief information including calling parameters and enriched motifs. 
}
\value{
No return values.
}
\seealso{
See Also as \code{\link{tfbs.scanTFsite}}
}
\examples{
#See example in tfbs.scanTFsite
}
\keyword{ print }
\keyword{Scanning}
