\name{tfbs.db-class}
\Rdversion{1.1}
\docType{class}
\alias{tfbs.db-class}

\title{Class \code{"tfbs.db"}}
\description{
Abstract class for motif dataset. The CisBP class is a son class of tfbs.db.
}
\section{Objects from the Class}{
Now code or function can be used to create this class.
}
\section{Slots}{
  \describe{
    \item{\code{species}:}{Species name. }
  }
}
\section{Methods}{
No methods defined with class "tfbs.db" in the signature.
}

\seealso{
}
\examples{
showClass("tfbs.db")
}
\keyword{classes}
