\name{write.starchbed}
\alias{write.starchbed}
\title{
Save BED file in a binary format
}
\description{
Save BED file in a binary format.
}
\usage{
write.starchbed(bed, filename)
}
\arguments{
  \item{bed}{Data frame with at least 3 columns.}
  \item{filename}{String indicating a filename.}
}
\details{
}
\value{
No return value.
}
\examples{
}
\keyword{internal}